Implement key management
We have three possible scenarios:
The user does not want to store the passphrase, then we have to ask him on every decryption attempt.
The user does not have a passphrase for the key, then just decrypt the password.
The user wants to store the passphrasem then decrypt the password using the stored passphrase.