Refactor entire code base to get rid of sandboxing
In the versions < 0.4 if Passafari, the host app was sandboxed.
The sandbox is removed now for multiple reasons.
First, due to the sandbox, the user had to import their private key to Passafari.
This required multiple steps, including exporting the private key to the ASCII armor format.
Additionally, the user had to specify the path to their password store.
And finally, Passafari had to handle the passphrase for the private key.
That was very ugly and I did not like it.
The main reason why the sandbox is gone now is because Passafari used a PGP Swift library,
which can not handle eliptic curve keys. That was the tipping point to remove the sandbox
to be able to let pass handle the password decryption.
This is more secure and more conventient for both, the developer and the user.