Search now excludes directories (fixes #17)
This change required a little more tweaking of the search.
The filesystem enumerator API for excluding directories operates on URLS
instead of Strings. This changed the entire path handling in the search
method, e.g., excluding the .git directory from the search did not
work anymore or the .gpg file extension was not removed properly.
The search methode was therefore more or less rewritten.