Introduce automated builds
In the past, Pass for macOS was built manually locally and the released by hand.
With this commit, this should change using automated builts with Github Actions.
The app is built, when a new tag v* is pushed.
It also creates a new release using Github Releases and updated the homebrew cask accordingly.