Fix bug preventing Pass for macOS running on older versions
Apple introduced Safari App Extension in macOS Sierra (10.12) together with
Safari 10. Therefire, I was aiming to support all macOS versions starting with
10.12 and Safari with 10. Unfortunately, I used the build target 10.14, which
prevented this. The build target was now changed to 10.12. This, however,
forced me to use NSLog instead of os.log for logging, as os.log is availble
starting with macOS 10.14.