Make global host app shortcut configurable (closes #38)
Until now the user had to use the default shortcut shift-ctrl-p to open the
host popup. This, however, interfered with some users workflow.
It is now possible to customize the global shortcut.
But this forced me to change quite a few things as I did not see any
convinient way doing this without some kind of preferences window.
So, this commit introduces a new preferences window.
However, to give the preferences some more meaning, I also moved all settings
regarding the update procedures to the preferences window.