Add function to determine if we are in Safari or not
As the goal is to unify the shortcuts, we have to find out if a) Safari is
front most and b) has at least one window. If both are true we can
safely say, that we have to auto-fill the password in Safari. If at least
one of them is false, we should show the status bar popover.
This is not done, yet. But this commit adds a function to determine if we
should show the popover or not.