Use host app's shortcut for password autofill
This commit adds the ability to use the host app's shortcut to autofill
passwords. This unifies the shortcuts, which is custimizable. To accomplish
this, the injected Javascript no longer listens for shortcuts but the host app
sends a signal to the extension using the Safari App Extension API. When the
extension receives the shortcut the remaining handling like searching for
passwords and auto-filling found passwords is mostly the same code as before.
There is now one major change in the behaviour. If there is only one password
found it will be filles in automatically. But if there are more than one
possible passwords, the extension shortcut will be opened showing the results.
If there is no Safari window in the foreground the status bar popover will be
shown as before.