Fix bug where ctrl-click was not handlet appropriately (fixes #42)
Pass for macOS's settings could only be accessed via right-clicking the status
bar item. However, this type of context menu should also be accessible with
ctrl-click, which was not possible unitl now. The odd thing is that I thought
that both are the same in macOS and the OS handles them the same way. The
problem is, however, that events are registered to the NSStatusItems button
property, which is a NSButton. NSButtons, however, are no receiving right-click
or ctrl-clicks anyways, as they are not designed to have a context menu.
Therefore, you have to implement different behaviours for left-click and
right-click on your own.