Add ability to cycle search results with tab
Until now, users had to cycle through the search results using the arrow key.
Further, there was no way to get back to the search field without using the
mouse. This was inconvenient and clumsy. This commit allows to cycle throught
the search results and the search field forward using tab and backward using
shift-tab.