Reset state after view disappears
Until now the state of the status bar popover remaind after the popover
disappeared. This was quite annoying as you had to change to the mouse, set the
cursor into the search field, remove the old search and start a new one. Now,
everytime the popover disappears, the state is reset.